CHORE: move files that not following the package rule
