+++ app/src/main/java/com/dearmyhealth/modules/Diet/fragment/DietTodayFragment.kt
@@ -1,4 +1,4 @@
-package com.dearmyhealth.modules.Diet.ui
+package com.dearmyhealth.modules.Diet.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,13 +6,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.dearmyhealth.databinding.FragmentDietTodayBinding
-import com.dearmyhealth.modules.Diet.DietDetailViewModel
+import com.dearmyhealth.modules.Diet.viewmodel.DietDetailViewModel
 
 
 class DietTodayFragment : Fragment() {
 
     private lateinit var binding: FragmentDietTodayBinding
-    private lateinit var viewModel: DietDetailViewModel
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
