+++ app/src/main/res/layout/fragment_diet_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".modules.Diet.ui.DietDetailFragment">
+    tools:context=".modules.Diet.fragment.DietDetailFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
