+++ app/src/main/res/layout/fragment_diet_stats.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".modules.Diet.ui.DietStatsFragment">
+    tools:context=".modules.Diet.fragment.DietStatsFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
