+++ app/src/main/res/layout/fragment_diet_today.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".modules.Diet.ui.DietTodayFragment">
+    tools:context=".modules.Diet.fragment.DietTodayFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
